requirements.txt
Django==5.0
djangorestframework==3.15.1
oracledb==2.1.0
reportlab==4.1.0